Simplify use-notifications
